adding refactored auth0 roles code
